pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "A simple and complete fastapi api exemple"
authors = ["Amos Pascal Amos <amos@kaanari.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
alembic = "^1.11.3"
uvicorn = "^0.23.2"
fastapi-debug-toolbar = "^0.5.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.3.0"}
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.20"}
ulid-py = "^1.1.0"
httpx = "^0.24.1"
humps = "^0.2.2"
psycopg2-binary = "^2.9.8"
asyncpg = "^0.28.0"
poethepoet = "^0.23.0"
pydantic-settings = "^2.0.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
sqlalchemyseed = "^2.0.0"
fire = "^0.5.0"
emails = "^0.6"
pika = "^1.3.2"
asgiref = "^3.7.2"
python-socketio = "^5.9.0"
cloudinary = "^1.36.0"
boto3 = "^1.28.62"
google-cloud-storage = "^2.11.0"

[tool.poetry.group.dev.dependencies]
locust = "^2.16.1"
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
mypy = "^1.5.1"
types-PyYAML = "^6.0.12.11"
types-toml = "^0.10.8.7"
autoflake = "^2.2.0"
pre-commit = "^3.3.3"
codespell = "^2.2.5"
bandit = "^1.7.5"
flake8 = "^6.1.0"
ruff = "^0.0.285"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
makemigrations = {"shell" = "alembic revision --autogenerate", help="Automatic creation of database migration files"}
migrate = {"shell" = "alembic upgrade heads", help="Database migration"}
dwngrade = {"shell" = "alembic downgrade -1", help="Downgrade the last migration"}
drop-tables = {"cmd" = "python -m seeder.run drop-tables", help="Drop all tables"}
seed = {"cmd" = "python -m seeder.run seed", help="Seed database"}
